@@ -13,17 +13,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.12",
-    "@angular/cdk": "^16.2.14",
-    "@angular/common": "^16.2.12",
-    "@angular/compiler": "^16.2.12",
-    "@angular/core": "^16.2.12",
-    "@angular/forms": "^16.2.12",
-    "@angular/material": "^16.2.14",
-    "@angular/platform-browser": "^16.2.12",
-    "@angular/platform-browser-dynamic": "^16.2.12",
-    "@angular/router": "^16.2.12",
-    "@angular/service-worker": "^16.2.12",
+    "@angular/animations": "^18.2.13",
+    "@angular/cdk": "^17.3.10",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/material": "^16.2.0",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
+    "@angular/service-worker": "^18.2.13",
     "@babel/plugin-transform-typescript": "^7.16.8",
     "aws-sdk": "^2.1348.0",
     "dayjs": "^1.10.7",
@@ -41,15 +41,15 @@
     "tslib": "^2.3.1",
     "uuid": "^8.3.2",
     "yarn": "^1.22.17",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
     "@angular-builders/jest": "^14.0.0",
     "@angular-devkit/build-angular": "^16.2.16",
     "@angular-eslint/schematics": "^20.1.1",
     "@angular/cli": "^16.2.16",
-    "@angular/compiler-cli": "^16.2.12",
-    "@angular/language-service": "^16.2.12",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/language-service": "^18.2.13",
     "@types/jasmine": "~3.5.14",
     "@types/jasminewd2": "~2.0.10",
     "@types/jest": "^28.1.2",
@@ -68,7 +68,7 @@
     "prettier-eslint": "^13.0.0",
     "protractor": "^7.0.0",
     "ts-node": "~8.3.0",
-    "typescript": "5.1.6"
+    "typescript": "5.5.4"
   },
   "jest-junit": {
     "outputDirectory": "./reports",
