@@ -12,7 +12,7 @@ import {RouterModule} from '@angular/router';
 import {SessionModule} from './features/session/session.module';
 import { Convertors } from './utilities/Convertors';
 import { ChordComponent } from './features/notation/chord/chord.component';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { ButtonModule } from 'primeng/button';
 import { InputTextModule } from 'primeng/inputtext';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
@@ -28,44 +28,37 @@ import { DialogModule } from "primeng/dialog";
 import { SongModule } from "./features/song/song.module";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    LandingPageComponent,
-    HeaderComponent,
-    FooterComponent,
-    NavigationComponent,
-    ChordComponent,
-    ExercisesComponent,
-    AuthComponent,
-    MetronomeComponent,
-    SongLibraryComponent,
-    UploadComponent
-  ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    AppRoutingModule,
-    SessionModule,
-    SongModule,
-    ReactiveFormsModule,
-    FormsModule,
-    ServiceWorkerModule.register("ngsw-worker.js", { enabled: environment.production }),
-    RouterModule,
-    HttpClientModule,
-    // PrimeNG
-    ButtonModule,
-    InputTextModule,
-    StyleClassModule,
-    MenuModule,
-    MenubarModule,
-    DialogModule
-  ],
-  providers: [Convertors],
-  schemas: [CUSTOM_ELEMENTS_SCHEMA],
-  exports: [
-    UploadComponent
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        LandingPageComponent,
+        HeaderComponent,
+        FooterComponent,
+        NavigationComponent,
+        ChordComponent,
+        ExercisesComponent,
+        AuthComponent,
+        MetronomeComponent,
+        SongLibraryComponent,
+        UploadComponent
+    ],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA],
+    exports: [
+        UploadComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        BrowserAnimationsModule,
+        AppRoutingModule,
+        SessionModule,
+        SongModule,
+        ReactiveFormsModule,
+        FormsModule,
+        ServiceWorkerModule.register("ngsw-worker.js", { enabled: environment.production }),
+        RouterModule,
+        // PrimeNG
+        ButtonModule,
+        InputTextModule,
+        StyleClassModule,
+        MenuModule,
+        MenubarModule,
+        DialogModule], providers: [Convertors, provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
